test: move chrome extensions spec to main runner
